requirements.txt
amqp==5.0.2
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
boto3==1.16.53
botocore==1.19.53
celery==5.0.5
cffi==1.14.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dj-database-url==0.5.0
Django==3.1.4
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-storages==1.11.1
Faker==5.0.2
gunicorn==20.0.4
importlib-metadata==2.1.1
jmespath==0.10.0
kombu==5.0.2
prompt-toolkit==3.0.8
psycopg2==2.8.6
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
rabbitmq==0.2.0
redis==3.5.3
s3transfer==0.3.4
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.7.4.3
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.0
